indexer/requirements.txt
psycopg2-binary
elasticsearch==8.11.1
tree-sitter
tree-sitter-python